@@ -1,16 +1,16 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-  ext.kotlin_version = '1.2.71'
+  ext.kotlin_version = '1.3.50'
   repositories {
     jcenter()
     google()
     maven { url "https://maven.fabric.io/public"}
   }
   dependencies {
-    classpath 'com.android.tools.build:gradle:3.2.1'
-    classpath 'com.google.gms:google-services:4.1.0'
-    classpath 'io.fabric.tools:gradle:1.25.4'
+    classpath 'com.android.tools.build:gradle:3.6.0-rc01'
+    classpath 'com.google.gms:google-services:4.3.3'
+    classpath 'io.fabric.tools:gradle:1.31.2'
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     classpath 'com.cookpad.android.licensetools:license-tools-plugin:1.5.0'
 
