@@ -1,11 +1,12 @@
 package com.futabooo.android.booklife
 
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import timber.log.Timber
 
 class InfiniteScrollListener(val func: () -> Unit,
-                             val layoutManager: LinearLayoutManager) : RecyclerView.OnScrollListener() {
+                             val layoutManager: LinearLayoutManager
+) : RecyclerView.OnScrollListener() {
 
   private var previousTotal = 0
   private var loading = true
