@@ -2,9 +2,9 @@ package com.futabooo.android.booklife.screen.option
 
 import android.content.Context
 import android.content.Intent
-import android.databinding.DataBindingUtil
+import androidx.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.futabooo.android.booklife.R
 import com.futabooo.android.booklife.databinding.ActivityOptionBinding
 
