@@ -1,6 +1,6 @@
 package com.futabooo.android.booklife
 
-import android.support.design.widget.TabLayout
+import com.google.android.material.tabs.TabLayout
 import android.view.View
 
 enum class MainBottomMenu constructor(val position: Int,
