@@ -2,10 +2,10 @@ package com.futabooo.android.booklife.screen.home
 
 import android.content.SharedPreferences
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.futabooo.android.booklife.BookLife
 import com.futabooo.android.booklife.databinding.FragmentHomeBinding
 import com.futabooo.android.booklife.model.HomeResource
@@ -13,14 +13,14 @@ import com.google.gson.Gson
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.schedulers.Schedulers
+import org.jsoup.Jsoup
+import retrofit2.Retrofit
+import timber.log.Timber
 import java.io.BufferedReader
 import java.io.InputStreamReader
 import java.text.SimpleDateFormat
 import java.util.Date
 import javax.inject.Inject
-import org.jsoup.Jsoup
-import retrofit2.Retrofit
-import timber.log.Timber
 
 class HomeFragment : Fragment() {
 
@@ -43,33 +43,43 @@ class HomeFragment : Fragment() {
     (activity?.application as BookLife).netComponent.inject(this)
   }
 
-  override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+  override fun onCreateView(
+    inflater: LayoutInflater,
+    container: ViewGroup?,
+    savedInstanceState: Bundle?
+  ): View? {
     binding = FragmentHomeBinding.inflate(inflater, container, false)
     return binding.root
   }
 
-  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+  override fun onViewCreated(
+    view: View,
+    savedInstanceState: Bundle?
+  ) {
     super.onViewCreated(view, savedInstanceState)
 
     val date = Date()
     val dateFormat = SimpleDateFormat("yyyy/MM")
     binding.fragmentHomeDate.text = dateFormat.format(date)
 
-    retrofit.create(HomeService::class.java).home()
+    retrofit.create(HomeService::class.java)
+        .home()
         .subscribeOn(Schedulers.io())
         .flatMap {
           val reader = BufferedReader(InputStreamReader(it.byteStream()))
-          val result = reader.readLines().filter(String::isNotBlank).toList()
+          val result = reader.readLines()
+              .filter(String::isNotBlank)
+              .toList()
 
-          val readingVolumes = Jsoup.parse(result.toString())
-              .select("div.home_index__userdata__main section.home_index__userdata__reading-volume")
-
-          val thisVolume = readingVolumes[0].select("ul span")
+          val thisVolume = Jsoup.parse(result.toString())
+              .select("div.stats__thismonth ul.thismonth__list li.list__item span.item__number")
           page = thisVolume[0].text()
-          volume = thisVolume[2].text()
-          pageParDay = thisVolume[4].text()
+          volume = thisVolume[1].text()
+          pageParDay = thisVolume[2].text()
 
-          val href = Jsoup.parse(result.toString()).select("div.home_index__userdata__side a").attr("href")
+          val href = Jsoup.parse(result.toString())
+              .select("div.bm-block-side__content dl.user-profiles dt.user-profiles__avatar a")
+              .attr("href")
           if (href.isNotBlank() && !sharedPreferences.contains("user_id")) {
             val userId = Integer.parseInt(href.substring(7))
             val editor = sharedPreferences.edit()
@@ -79,7 +89,8 @@ class HomeFragment : Fragment() {
 
           val csrfToken = Jsoup.parse(result.toString()).select("meta[name=csrf-token]")[0].attr("content")
 
-          retrofit.create(HomeService::class.java).getJson(csrfToken, 0, 10)
+          retrofit.create(HomeService::class.java)
+              .getJson(csrfToken, 0, 10)
         }
         .observeOn(AndroidSchedulers.mainThread())
         .subscribeBy(
@@ -99,7 +110,8 @@ class HomeFragment : Fragment() {
               binding.readingPageParDayCurrentMonth.text = pageParDay
             },
             onError = {
-              Timber.e(it, it.message) }
+              Timber.e(it, it.message)
+            }
         )
   }
 }
