@@ -5,23 +5,23 @@ import android.animation.AnimatorListenerAdapter
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
-import android.databinding.DataBindingUtil
 import android.net.Uri
 import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AppCompatActivity
 import android.text.TextUtils
 import android.view.View
 import android.view.inputmethod.EditorInfo
 import android.view.inputmethod.InputMethodManager
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
 import com.futabooo.android.booklife.BookLife
 import com.futabooo.android.booklife.R
 import com.futabooo.android.booklife.databinding.ActivityLoginBinding
 import com.futabooo.android.booklife.screen.MainActivity
+import com.google.android.material.snackbar.Snackbar
 import com.kazakago.cryptore.Cryptore
-import javax.inject.Inject
 import retrofit2.Retrofit
+import javax.inject.Inject
 
 class LoginActivity : AppCompatActivity(), LoginPresenter.Contract {
 
